+++ extensions/azure/cosmos/policy-store-cosmos/src/main/resources/META-INF/services/org.eclipse.dataspaceconnector.spi.system.ServiceExtension
@@ -12,5 +12,5 @@
 #
 #
 
-org.eclipse.dataspaceconnector.contract.definition.store.CosmosPolicyStoreExtension
+org.eclipse.dataspaceconnector.cosmos.policy.store.CosmosPolicyStoreExtension
