+++ extensions/azure/cosmos/policy-store-cosmos/src/test/java/org/eclipse/dataspaceconnector/cosmos/policy/store/model/PolicyDocumentSerializationTest.java
@@ -12,16 +12,17 @@
  *
  */
 
-package org.eclipse.dataspaceconnector.contract.definition.store.model;
+package org.eclipse.dataspaceconnector.cosmos.policy.store.model;
 
+import org.eclipse.dataspaceconnector.cosmos.policy.store.PolicyDocument;
 import org.eclipse.dataspaceconnector.spi.types.TypeManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.eclipse.dataspaceconnector.contract.definition.store.TestFunctions.generateDefinition;
+import static org.eclipse.dataspaceconnector.cosmos.policy.store.TestFunctions.generatePolicy;
 
-class ContractDefinitionDocumentSerializationTest {
+class PolicyDocumentSerializationTest {
 
     private TypeManager typeManager;
 
@@ -29,33 +30,31 @@ class ContractDefinitionDocumentSerializationTest {
     @BeforeEach
     void setup() {
         typeManager = new TypeManager();
-        typeManager.registerTypes(ContractDefinitionDocument.class, ContractDefinitionDocument.class);
     }
 
     @Test
     void testSerialization() {
-        var def = generateDefinition();
+        var def = generatePolicy();
         var pk = "test-part-key";
 
-        var document = new ContractDefinitionDocument(def, pk);
+        var document = new PolicyDocument(def, pk);
 
         String s = typeManager.writeValueAsString(document);
 
         assertThat(s).isNotNull()
-                .contains("\"selectorExpression\":{\"criteria\":[{\"left\":\"somekey\",\"op\":\"=\",\"right\":\"someval\"}]}}")
                 .contains("wrappedInstance")
-                .contains("\"id\":\"" + def.getId() + "\"")
+                .contains("\"id\":\"" + def.getUid() + "\"")
                 .contains("\"partitionKey\":\"" + pk + "\"");
     }
 
     @Test
     void testDeserialization() {
-        var def = generateDefinition();
+        var def = generatePolicy();
 
-        var document = new ContractDefinitionDocument(def, "test-part-key");
+        var document = new PolicyDocument(def, "test-part-key");
         String json = typeManager.writeValueAsString(document);
 
-        var transferProcessDeserialized = typeManager.readValue(json, ContractDefinitionDocument.class);
+        var transferProcessDeserialized = typeManager.readValue(json, PolicyDocument.class);
         assertThat(transferProcessDeserialized).usingRecursiveComparison().isEqualTo(document);
     }
 }
